Add method for finishing the task
Added calculation for duration that rounds to two decimal points hours. It checks if the task_id is in data json format, iand if the task is in the database. Then it updates this task in the database and it effects also getting all the tasks for user. It does not delete it to be in future statistics.